Validaciones y Documentación

+++ Api/Controllers/Inventario/ArticulosController.cs
@@ -1,4 +1,6 @@
 ﻿using Api.Controllers;
+using Api.Errors;
+using Api.Models.General;
 using Api.Models.Inventario;
 using Dapper;
 using Microsoft.AspNetCore.Mvc;
@@ -8,7 +10,7 @@ using System.Data.SqlClient;
 namespace Api.Controllers.Inventario
 {
     [ApiController]
-    [Route("[controller]")]
+    [Route("api/[controller]")]
     public class ArticulosController : ControllerBase
     {
         private readonly IConfiguration _configuration;
@@ -22,8 +24,10 @@ namespace Api.Controllers.Inventario
         }
 
         /// <summary>
-        /// Obtiene todos los articulos.
+        /// Obtiene todos los artículos.
         /// </summary>
+        /// <returns>Una lista de los artículos en la base de datos.</returns>
+        /// <exception cref="ApiException">Se lanza si no se encuentran artículos registrados.</exception>
         [HttpGet]
         public async Task<ActionResult<List<Articulos>>> GetAllArticulos()
         {
@@ -32,20 +36,34 @@ namespace Api.Controllers.Inventario
                 using (var connection = new SqlConnection(_configuration.GetConnectionString("DefaultConnection")))
                 {
                     var articulos = await connection.QueryAsync<Articulos>("select * from tblarticulos");
-                    return Ok(articulos.ToList());
+
+                    if (articulos.Any())
+                    {
+                        return Ok(articulos.ToList());
+                    }
+                    else
+                    {
+                        throw new ApiException(404, "No hay artículos registrados.");
+                    }
                 }
+            }
+            catch (ApiException)
+            {
+                throw;
             }
             catch (Exception ex)
             {
                 _logger.LogError(ex, "Error al obtener los artículos.");
-                return StatusCode(500, "Error al obtener los artículos.");
+                throw new ApiException(500, "Ha ocurrido un error interno al obtener los artículos. Detalle: " + ex.Message);
             }
         }
 
         /// <summary>
-        /// Obtiene un articulo por su ID.
+        /// Obtiene un artículo por su ID.
         /// </summary>
-        /// <param name="codigo">Codigo del articulo.</param>
+        /// <param name="codigo">Codigo del artículo.</param>
+        /// <returns>El artículo específico que se consulta por el codigo.</returns>
+        /// <exception cref="ApiException">Se lanza si no se encuentra el artículo correspondiente al código <paramref name="codigo"/> del artículo.</exception>
         [HttpGet("{codigo}")]
         public async Task<ActionResult<Articulos>> GetArticulo(string codigo)
         {
@@ -55,13 +73,25 @@ namespace Api.Controllers.Inventario
                 {
                     var articulo = await connection.QueryAsync<Articulos>("select * from tblarticulos where codigo = @Codigo",
                         new { Codigo = codigo });
-                    return Ok(articulo.First());
+
+                    if (articulo.Any())
+                    {
+                        return Ok(articulo.First());
+                    }
+                    else
+                    {
+                        throw new ApiException(404, "El artículo con el código '" + codigo + "' no está registrado.");
+                    }
                 }
+            }
+            catch (ApiException)
+            {
+                throw;
             }
             catch (Exception ex)
             {
                 _logger.LogError(ex, "Error al obtener el artículo " + codigo);
-                return StatusCode(500, "Error al obtener el artículo.");
+                throw new ApiException(500, "Ha ocurrido un error interno al obtener el artículo. Detalle: " + ex.Message);
             }
         }
 
@@ -69,7 +99,8 @@ namespace Api.Controllers.Inventario
         /// Crea un nuevo articulo.
         /// </summary>
         /// <param name="articulo">Información del articulo a crear.</param>
-        /// <returns></returns>
+        /// <returns>Los artículos que devuelve el método SelectAllArticulos.</returns>
+        /// <exception cref="ApiException">Se lanza si ya existe el artículo correspondiente al código del artículo <paramref name="articulo"/>.</exception>
         [HttpPost]
         public async Task<ActionResult<List<Articulos>>> CreateArticulo(Articulos articulo)
         {
@@ -77,43 +108,71 @@ namespace Api.Controllers.Inventario
             {
                 using (var connection = new SqlConnection(_configuration.GetConnectionString("DefaultConnection")))
                 {
-                    await connection.ExecuteAsync("INSERT INTO TblArticulos (Codigo, Stock, Nombre, Descripcion, PrecioCompra, PrecioVenta, CategoriaId, UnidadMedidaId, EstadoId) VALUES (@Codigo, @Stock, @Nombre, @Descripcion, @PrecioCompra, @PrecioVenta, @CategoriaId, @UnidadMedidaId, @EstadoId)", articulo);
-                    return Ok(await SelectAllArticulos(connection));
+                    var result = await SelectArticuloId(connection, articulo.Codigo);
+                    if (!result.Any())
+                    {
+                        await connection.ExecuteAsync("insert into TblArticulos (Codigo, Stock, Nombre, Descripcion, PrecioCompra, PrecioVenta, CategoriaId, UnidadMedidaId, EstadoId) VALUES (@Codigo, @Stock, @Nombre, @Descripcion, @PrecioCompra, @PrecioVenta, @CategoriaId, @UnidadMedidaId, @EstadoId)", articulo);
+                        return Ok(await SelectAllArticulos(connection));
+                    }
+                    else
+                    {
+                        throw new ApiException(409, "El artículo con el código '" + articulo.Codigo + "' ya está registrado.");
+                    }
                 }
+            }
+            catch (ApiException)
+            {
+                throw;
             }
             catch (Exception ex)
             {
                 _logger.LogError(ex, "Error al crear el artículo.");
-                return StatusCode(500, "Error al crear el artículo.");
+                throw new ApiException(500, "Ha ocurrido un error interno al crear el artículo. Detalle: " + ex.Message);
             }
         }
 
-        // <summary>
+        /// <summary>
         /// Actualiza un articulo.
         /// </summary>
         /// <param name="articulo">Información del articulo a actualizar.</param>
+        /// <returns>Los artículos que devuelve el método SelectAllArticulos.</returns>
+        /// <exception cref="ApiException">Se lanza si no se encuentra el artículo correspondiente al código del artículo <paramref name="articulo"/>.</exception>
         [HttpPut]
-        public async Task<ActionResult<List<Articulos>>> Updatearticulo(Articulos articulo)
+        public async Task<ActionResult<List<Articulos>>> UpdateArticulo(Articulos articulo)
         {
             try
             {
                 using (var connection = new SqlConnection(_configuration.GetConnectionString("DefaultConnection")))
                 {
-                    await connection.ExecuteAsync("UPDATE TblArticulos SET Codigo = @Codigo, Stock = @Stock, Nombre = @Nombre, Descripcion = @Descripcion, PrecioCompra = @PrecioCompra, PrecioVenta = @PrecioVenta, CategoriaId = @CategoriaId, UnidadMedidaId = @UnidadMedidaId, EstadoId = @EstadoId WHERE Id = @Id", articulo);
-                    return Ok(await SelectAllArticulos(connection));
+                    var result = await SelectArticuloId(connection, articulo.Codigo);
+                    if (result.Any())
+                    {
+                        await connection.ExecuteAsync("UPDATE TblArticulos SET Codigo = @Codigo, Stock = @Stock, Nombre = @Nombre, Descripcion = @Descripcion, PrecioCompra = @PrecioCompra, PrecioVenta = @PrecioVenta, CategoriaId = @CategoriaId, UnidadMedidaId = @UnidadMedidaId, EstadoId = @EstadoId WHERE Id = @Id", articulo);
+                        return Ok(await SelectAllArticulos(connection));
+                    }
+                    else
+                    {
+                        throw new ApiException(404, "El artículo con el código '" + articulo.Codigo + "' no está registrado.");
+                    }
                 }
+            }
+            catch (ApiException)
+            {
+                throw;
             }
             catch (Exception ex)
             {
                 _logger.LogError(ex, "Error al actualizar el artículo " + articulo.Id);
-                return StatusCode(500, "Error al actualizar el artículo.");
+                throw new ApiException(500, "Ha ocurrido un error interno al actualizar el artículo. Detalle: " + ex.Message);
             }
         }
 
-        // <summary>
+        /// <summary>
         /// Elimina un articulo.
         /// </summary>
         /// <param name="articulo">Información del articulo a eliminar.</param>
+        /// <returns>Los artículos que devuelve el método SelectAllArticulos.</returns>
+        /// <exception cref="ApiException">Se lanza si no se encuentra el artículo correspondiente al código del artículo <paramref name="articulo"/>.</exception>
         [HttpDelete]
         public async Task<ActionResult<List<Articulos>>> Deletearticulo(Articulos articulo)
         {
@@ -121,14 +180,26 @@ namespace Api.Controllers.Inventario
             {
                 using (var connection = new SqlConnection(_configuration.GetConnectionString("DefaultConnection")))
                 {
-                    await connection.ExecuteAsync("delete from tblarticulos where id = @id", articulo);
-                    return Ok(await SelectAllArticulos(connection));
+                    var result = await SelectArticuloId(connection, articulo.Codigo);
+                    if (result.Any())
+                    {
+                        await connection.ExecuteAsync("delete from tblarticulos where id = @id", articulo);
+                        return Ok(await SelectAllArticulos(connection));
+                    }
+                    else
+                    {
+                        throw new ApiException(404, "El artículo con el código '" + articulo.Codigo + "' no está registrado.");
+                    }
                 }
+            }
+            catch (ApiException)
+            {
+                throw;
             }
             catch (Exception ex)
             {
                 _logger.LogError(ex, "Error al eliminar el artículo " + articulo.Id);
-                return StatusCode(500, "Error al eliminar el artículo.");
+                throw new ApiException(500, "Ha ocurrido un error interno al eliminar el artículo. Detalle: " + ex.Message);
             }
         }
 
@@ -143,23 +214,15 @@ namespace Api.Controllers.Inventario
         }
 
         /// <summary>
-        /// Selecciona todos los articulos.
+        /// Selecciona un artículo por identificador.
         /// </summary>
         /// <param name="connection">Cadena de conexión a la base de datos.</param>
-        /// <returns>Todos los articulos de la tabla tblarticulos.</returns>
-        private static async Task<bool> ArticuloExiste(SqlConnection connection, Articulos articulo)
+        /// <param name="codigo">Identificador del artículo.</param>
+        /// <returns>El artículo que se está consultando.</returns>
+        private static async Task<IEnumerable<Articulos>> SelectArticuloId (SqlConnection connection, string codigo)
         {
-            var respuesta = await connection.QueryAsync<Articulos>("select * from tblarticulos where codigo = @codigo ", 
-                new { codigo = articulo.Codigo });
-
-            if (respuesta == null)
-            {
-                return true;
-            }
-            else
-            {
-                return false;
-            }
+            return await connection.QueryAsync<Articulos>("select * from tblarticulos where codigo = @Codigo", 
+                new { Codigo = codigo });
         }
     }
 }
