+++ Api/Controllers/Inventario/CategoriasController.cs
@@ -1,4 +1,6 @@
-﻿using Api.Models.Inventario;
+﻿using Api.Errors;
+using Api.Models.General;
+using Api.Models.Inventario;
 using Dapper;
 using Microsoft.AspNetCore.Mvc;
 using System.Data.SqlClient;
@@ -6,7 +8,7 @@ using System.Data.SqlClient;
 namespace Api.Controllers.Inventario
 {
     [ApiController]
-    [Route("[controller]")]
+    [Route("api/[controller]")]
     public class CategoriasController : Controller
     {
         private readonly IConfiguration _configuration;
@@ -20,9 +22,11 @@ namespace Api.Controllers.Inventario
         }
 
         /// <summary>
-        /// Obtiene todos los categorias.
+        /// Obtiene todas las categorias.
         /// </summary>
-        [HttpGet(Name = "Getcategorias")]
+        /// <returns>Una lista de las categorías en la base de datos.</returns>
+        /// <exception cref="ApiException">Se lanza si no se encuentran categorías registradas.</exception>
+        [HttpGet]
         public async Task<ActionResult<List<Categorias>>> GetAllCategorias()
         {
             try
@@ -30,20 +34,34 @@ namespace Api.Controllers.Inventario
                 using (var connection = new SqlConnection(_configuration.GetConnectionString("DefaultConnection")))
                 {
                     var categorias = await connection.QueryAsync<Categorias>("select * from tblcategorias");
-                    return Ok(categorias.ToList());
+                    
+                    if(categorias.Any())
+                    {
+                        return Ok(categorias.ToList());
+                    }
+                    else
+                    {
+                        throw new ApiException(404, "No hay categorías registradas.");
+                    }
                 }
+            }
+            catch (ApiException)
+            {
+                throw;
             }
             catch (Exception ex)
             {
                 _logger.LogError(ex, "Error al obtener categorias.");
-                return StatusCode(500, "Error al obtener categorias.");
+                throw new ApiException(500, "Ha ocurrido un error interno al obtener las categorias. Detalle: " + ex.Message);
             }
         }
 
         /// <summary>
-        /// Obtiene un categoria por su ID.
+        /// Obtiene un categoría por su ID.
         /// </summary>
-        /// <param name="categoriaId">ID del categoria.</param>
+        /// <param name="categoriaId">Id de la categoría.</param>
+        /// <returns>La categoría que corresponde al ID <paramref name="categoriaId"/> en la base de datos.</returns>
+        /// <exception cref="ApiException">Se lanza si no se encuentra la categoría correspondiente al ID <paramref name="categoriaId"/>.</exception>
         [HttpGet("{categoriaId:int}")]
         public async Task<ActionResult<Categorias>> GetCategoria(int categoriaId)
         {
@@ -53,20 +71,34 @@ namespace Api.Controllers.Inventario
                 {
                     var categoria = await connection.QueryAsync<Categorias>("select * from tblcategorias where id = @Id",
                         new { Id = categoriaId });
-                    return Ok(categoria.First());
+                    
+                    if (categoria.Any())
+                    {
+                        return Ok(categoria.First());
+                    }
+                    else
+                    {
+                        throw new ApiException(404, "La categoría con el id '" + categoriaId + "' no está registrada.");
+                    }
                 }
+            }
+            catch (ApiException)
+            {
+                throw;
             }
             catch (Exception ex)
             {
-                _logger.LogError(ex, "Error al obtener el categoria " + categoriaId);
-                return StatusCode(500, "Error al obtener el categoria.");
+                _logger.LogError(ex, "Error al obtener el categoría " + categoriaId);
+                throw new ApiException(500, "Ha ocurrido un error interno al obtener la categoría. Detalle: " + ex.Message);
             }
         }
 
-        // <summary>
-        /// Crea un nuevo categoria.
+        /// <summary>
+        /// Crea una nueva categoría.
         /// </summary>
-        /// <param name="categoria">Información del categoria a crear.</param>
+        /// <param name="categoria">Información de la categoría a crear.</param>
+        /// <returns>Las categorías en la base de datos. Consulte <see cref="SelectAllCategorias"/> para obtener más detalles.</returns>
+        /// <exception cref="ApiException">Se lanza si ya existe la categoría correspondiente al ID de la categoría <paramref name="categoria"/>.</exception>
         [HttpPost]
         public async Task<ActionResult<List<Categorias>>> CreateCategoria(Categorias categoria)
         {
@@ -74,21 +106,35 @@ namespace Api.Controllers.Inventario
             {
                 using (var connection = new SqlConnection(_configuration.GetConnectionString("DefaultConnection")))
                 {
-                    await connection.ExecuteAsync("insert into tblcategorias (nombre, descripcion) values (@nombre, @descripcion)", categoria);
-                    return Ok(await SelectAllCategorias(connection));
+                    var result = await SelectCategoriaId(connection, categoria.Id);
+                    if (!result.Any())
+                    {
+                        await connection.ExecuteAsync("insert into tblcategorias (nombre, descripcion) values (@nombre, @descripcion)", categoria);
+                        return Ok(await SelectAllCategorias(connection));
+                    }
+                    else
+                    {
+                        throw new ApiException(409, "La categoría con el Id '" + categoria.Id + "' ya está registrada.");
+                    }
                 }
+            }
+            catch (ApiException)
+            {
+                throw;
             }
             catch (Exception ex)
             {
-                _logger.LogError(ex, "Error al crear el categoria.");
-                return StatusCode(500, "Error al crear el categoria.");
+                _logger.LogError(ex, "Error al crear la categoría.");
+                throw new ApiException(500, "Ha ocurrido un error interno al crear la categoría. Detalle: " + ex.Message);
             }
         }
 
-        // <summary>
-        /// Actualiza un categoria.
+        /// <summary>
+        /// Actualiza un categoría.
         /// </summary>
-        /// <param name="categoria">Información del categoria a actualizar.</param>
+        /// <param name="categoria">Información de la categoría a actualizar.</param>
+        /// <returns>Las categorías en la base de datos. Consulte <see cref="SelectAllCategorias"/> para obtener más detalles.</returns>
+        /// <exception cref="ApiException">Se lanza si no se encuentra la categoría correspondiente al ID de la categoría <paramref name="categoria"/>.</exception>
         [HttpPut]
         public async Task<ActionResult<List<Categorias>>> UpdateCategoria(Categorias categoria)
         {
@@ -96,21 +142,35 @@ namespace Api.Controllers.Inventario
             {
                 using (var connection = new SqlConnection(_configuration.GetConnectionString("DefaultConnection")))
                 {
-                    await connection.ExecuteAsync("update tblcategorias set nombre = @nombre, descripcion = @descripcion where id = @id", categoria);
-                    return Ok(await SelectAllCategorias(connection));
+                    var result = await SelectCategoriaId(connection, categoria.Id);
+                    if (result.Any())
+                    {
+                        await connection.ExecuteAsync("update tblcategorias set nombre = @nombre, descripcion = @descripcion where id = @id", categoria);
+                        return Ok(await SelectAllCategorias(connection));
+                    }
+                    else
+                    {
+                        throw new ApiException(404, "La categoría con el ID '" + categoria.Id + "' no está registrada.");
+                    } 
                 }
+            }
+            catch (ApiException)
+            {
+                throw;
             }
             catch (Exception ex)
             {
-                _logger.LogError(ex, "Error al actualizar el categoria " + categoria.Id);
-                return StatusCode(500, "Error al actualizar el categoria.");
+                _logger.LogError(ex, "Error al actualizar la categoría " + categoria.Id);
+                throw new ApiException(500, "Ha ocurrido un error interno al actualizar la categoría. Detalle: " + ex.Message);
             }
         }
 
-        // <summary>
-        /// Elimina un categoria.
+        /// <summary>
+        /// Elimina una categoría.
         /// </summary>
-        /// <param name="categoria">Información del categoria a eliminar.</param>
+        /// <param name="categoria">Información de la categoría a eliminar.</param>
+        /// <returns>Las categorías en la base de datos. Consulte <see cref="SelectAllCategorias"/> para obtener más detalles.</returns>
+        /// <exception cref="ApiException">Se lanza si no se encuentra la categoría correspondiente al ID de la categoría <paramref name="categoria"/>.</exception>
         [HttpDelete]
         public async Task<ActionResult<List<Categorias>>> DeleteCategoria(Categorias categoria)
         {
@@ -118,25 +178,48 @@ namespace Api.Controllers.Inventario
             {
                 using (var connection = new SqlConnection(_configuration.GetConnectionString("DefaultConnection")))
                 {
-                    await connection.ExecuteAsync("delete from tblcategorias where id = @id", categoria);
-                    return Ok(await SelectAllCategorias(connection));
+                    var result = await SelectCategoriaId(connection, categoria.Id);
+                    if (result.Any())
+                    {
+                        await connection.ExecuteAsync("delete from tblcategorias where id = @id", categoria);
+                        return Ok(await SelectAllCategorias(connection));
+                    }
+                    else
+                    {
+                        throw new ApiException(404, "La categoría con el ID '" + categoria.Id + "' no está registrada.");
+                    }
                 }
+            }
+            catch (ApiException)
+            {
+                throw;
             }
             catch (Exception ex)
             {
-                _logger.LogError(ex, "Error al eliminar el categoria " + categoria.Id);
-                return StatusCode(500, "Error al eliminar el categoria.");
+                _logger.LogError(ex, "Error al eliminar la categoría " + categoria.Id);
+                throw new ApiException(500, "Ha ocurrido un error interno al eliminar la categoría. Detalle: " + ex.Message);
             }
         }
 
         /// <summary>
-        /// Selecciona todos los categorias.
+        /// Selecciona todas las categorías.
         /// </summary>
         /// <param name="connection">Cadena de conexión a la base de datos.</param>
-        /// <returns>Todos los categorias de la tabla tblcategorias.</returns>
+        /// <returns>Todas las categorías de la tabla tblcategorias.</returns>
         private static async Task<IEnumerable<Categorias>> SelectAllCategorias(SqlConnection connection)
         {
             return await connection.QueryAsync<Categorias>("select * from tblcategorias");
+        }
+
+        /// <summary>
+        /// Consulta una categoría por su identificador.
+        /// </summary>
+        /// <param name="connection">Cadena de conexión a la base de datos.</param>
+        /// <param name="id">Identificador de la categoría.</param>
+        /// <returns>La categoría que corresponde al ID <paramref name="id"/> en la base de datos.</returns>
+        private static async Task<IEnumerable<Categorias>> SelectCategoriaId(SqlConnection connection, int id)
+        {
+            return await connection.QueryAsync<Categorias>("select * from tblcategorias where id = @Id", new { id = id });
         }
     }
 }
